First meaningful commit.  Please still ignore it.
This commit puts the initial code up on the web.  It's still pretty
gnarly, and contains code I've since refactored, removed, etc.  Part of
the problem is that I develop on a Surface RT, and there's no ARM
version of windows Git.  So I've got a boat load of changes in engine in
particular, but also in other files.  Those changes (still a little
hairy) improve this thing a LOT.  Surface is at my office, though, and
here I sit in the suburbs.

Awesome commit message, bro.  Commit it again :|