Vast improvement - for timing, clearer debugging.
Changes:

1) engine.debug.js - Added debugging caveats for writing certain
debugging items.
2) index.html - fixed resume button on Zats, uncommented debug file
inclusion
3) engine.js - fixed timing problems, and in so doing, removed the need
for repeated input timers.  Broke flying mode in the process -- can't
win them all.  Now we are starting from more of a solid foundation
4) index.css - changed some sizes on header and command history

This puts the input and game flow on a much more stable, predictable,
and understandable track.  This version can be recommended for testing
(to get the general idea of how the program works) and to some extent
for observing the code to grok general functionality.

Some previously working things are broken, and there are still some
serious refactorings needed, but there's better commenting and also a
better foundation.