Engine Core separation, cmdhist changes, and more
Changed the way the Command History works - bottom up, like the
original! And awesome because it was much easier than expected.
Reformatted the map data files to visually aid my development/edit of
the test maps.
Pulled a bunch of uiir.engine into core and added some notes/comments
for my later awareness.
Added some functionality to player allowing returns of 0-padded versions
of some of its user-facing values.
Moved the game into uiirmain.html, and turned index into a menu --
useful only for my own debugging, and I've removed the other files to
which it, for a brief time, linked.
Modified css, moving the ui of the game and its assets a LOT closer to
matching the original.
Added an "action plans" text so I wouldn't forget what I was planning to
do next.  I want to replace this with a website.
Added a "bug list" text so I could remember what's glitchy and what I'm
thinking about doing about them, if anything.  I want to replace this
with a website.
Saved a copy of the old uiir drawing code so I had something to look at
while I wrote more amazing code elsewhere -- or, you know, when I moved
it.  I forget. This'll probably be deleted with the next push.
Wrote code for the switch over to uiir command -- I'm not sure how much
of it I'll be using.  That's my next official effort, though I'm
sidelining it for a bit to add the terrain class in, and try and nail
down a map class.
Wrote new drawing code.  It's cleaner, smaller, and takes advantage of
some js features I wasn't using before. This (draw2.js) will probably be
renamed in the next push.
Added a copy of the old uiir.engine code, prior to moving stuff around.
I think this was a reference item, and it'll almost certainly be deleted
with the next push.
Added modes for game, loading, and zats -- all of which are utilized now
-- for better separation of logic and an easier-to read codebase.
Added all Thumbs.db files to .hgignore.